#51 refactor: logout 메서드를 AuthService로 이동
