+++ src/main/java/kr/codesquad/secondhand/infrastructure/properties/RedisProperties.java
@@ -3,14 +3,23 @@ package kr.codesquad.secondhand.infrastructure.properties;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.stereotype.Component;
 
 @Setter
 @Getter
-@ConfigurationProperties(prefix = "redis")
 @Component
+@ConfigurationProperties(prefix = "redis")
 public class RedisProperties {
 
     private int port;
     private String host;
+
+    public int getPort() {
+        return this.port;
+    }
+
+    public String getHost() {
+        return this.host;
+    }
 }
