+++ src/main/resources/application-redis.yml
@@ -1,3 +1,3 @@
 redis:
-    host: 127.0.0.1
+    host: localhost
     port: 6379
