+++ src/main/resources/application-test.yml
@@ -39,5 +39,5 @@ oauth2:
     user-info-url: user-info-url
 
 redis:
-  host: 127.0.0.1
+  host: localhost
   port: 6379
